backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
langchain-community
langchain-core
pandas
numpy
langchain-openai
langchain-chroma
